Added script to place and rotate virtual chair
This is working, but is a bit finicky. The app will most likely register 1 finger first and move the chair before the second finger is registrered and the rotation can start. The chair also doesn't remember the rotation, so one has to rotate the chair each time the chair is moved to a new place.