Added init scene and more participant number stuff
Added an init scene which handles global variables and other start up
stuff, such as making a participant number that will last throughout the
session. The participant number is also added to the finish screen, to
make it easier for the participant to remember the number once they will
fill out the post-testing questionnaire. Closes issue #27 and #26.