Various instruction changes
In Task 1 dialog box, it now tells the participant to use slow and
controlled movements (issue #25) and at the end it tells the participant
to try to map most of the floor in the room (issue #24).

Fixed issue #29, by making the Task 1 dialog box be on when the AREnv
scene is initiated.